SmartRV/pyproject.toml

[tool.poetry]
name = "WinnConnect Coach"
version = "0.1.0"
description = ""
authors = ["Dominique Parolin <d.parolin@gmx.net>"]

[tool.poetry.dependencies]
python3 = "^3.10"
aenum = "^3.1.11"
anyio = "^3.6.1"
argparse-addons = "^0.8.0"
asn1crypto = "^1.5.1"
atomicwrites = "^1.4.0"
attrs = "^21.4.0"
azure-core = "^1.26.4"
azure-identity = "^1.13.0"
azure-iot-device = "^2.12.0"
azure-storage-blob = "^12.16.0"
bcrypt = "^3.2.0"
beautifulsoup4 = "^4.11.1"
betamax = "^0.8.1"
bitstring = "^3.1.9"
bitstruct = "^8.14.0"
cachecontrol = "^0.12.10"
cachy = "^0.3.0"
python-can = "^4.0.0"
cantools = "^39.0.0"
certifi = "^2021.10.8"
cffi = "^1.15.0"
charset-normalizer = "^2.0.12"
cleo = "^0.8.1"
click = "^8.1.2"
clikit = "^0.6.2"
cmd2 = "^2.4.1"
colorama = "^0.4.4"
crashtest = "^0.3.1"
crccheck = "^1.3.0"
cryptography = "^36.0.2"
cryptography-vectors = "^36.0.2"
#dbus = "^1.2.18"
deprecation = "^2.1.0"
dictdiffer = "^0.9.0"
diskcache = "^5.4.0"
distro = "^1.7.0"
ecdsa = "^0.17.0"
fastapi = "^0.83.0"
filelock = "^3.8.0"
gmpy2 = "^2.1.2"
h11 = "^0.13.0"
html5lib = "^1.1"
hypothesis = "^6.39.5"
idna = "^3.3"
importlib-metadata = "^4.11.3"
iniconfig = "^1.1.1"
iniparse = "^0.5"
iso8601 = "^1.0.2"
isodate = "^0.6.1"
janus = "^1.0.0"
jinja2 = "^3.1.1"
lockfile = "^0.12.2"
lxml = "^4.8.0"
markupsafe = "^2.1.1"
mccabe = "^0.7.0"
mock = "^4.0.3"
more-itertools = "^8.12.0"
msgpack = "^1.0.3"
markdown2 = "^1.1.0"
packaging = "^21.3"
paho-mqtt = "^1.6.1"
passlib = "^1.7.4"
pastel = "^0.2.1"
pathlib2 = "^2.3.7"
pbr = "^5.8.1"
pep8 = "^1.7.1"
pexpect = "^4.8.0"
pip = "^22.0.3"
pkginfo = "^1.8.3"
pluggy = "^1.0.0"
ply = "^3.11"
pretend = "^1.0.9"
prettytable = "^3.1.1"
psutil = "^5.9.0"
ptyprocess = "^0.7.0"
py = "^1.11.0"
pyasn1 = "^0.4.8"
pybind11 = "^2.8.1"
pycairo = "^1.21.0"
pycparser = "^2.21"
pydantic = "^1.10.7"
pyflakes = "^2.4.0"
pygobject = "^3.42.0"
pylev = "^1.4.0"
pyopenssl = "^22.0.0"
pyparsing = "^3.0.7"
pyperclip = "^1.8.2"
pypng = "^0.0.20"
pyqrcode = "^1.2.1"
pyserial = "^3.5"
pysocks = "^1.7.1"
pytest = "^7.1.1"
pytest-lazy-fixture = "^0.6.3"
pytest-subtests = "^0.7.0"
python-jose = "^3.3.0"
pytz = "^2022.1"
requests = "^2.28.1"
requests-toolbelt = "^0.9.1"
requests-unixsocket = "^0.3.0"
rsa = "^4.8"
secretstorage = "^3.3.3"
setproctitle = "^1.3.2"
setuptools = "^59.5.0"
#setuptools-rust-native = "^1.1.2"
shellingham = "^1.5.0"
six = "^1.16.0"
sniffio = "^1.2.0"
sortedcontainers = "^2.4.0"
soupsieve = "^2.3.1"
starlette = "^0.19.1"
textparser = "^0.24.0"
toml = "^0.10.2"
tomli = "^2.0.1"
tomlkit = "^0.11.4"
typing-extensions = "^4.3.0"
urllib3 = "^1.26.9"
uvicorn = "^0.18.2"
wcwidth = "^0.2.5"
webencodings = "^0.5.1"
websockets = "^10.2"
wrapt = "^1.14.0"
zipp = "^3.7.0"
jsonschema = "^4.19"

# Test Dependencies

[tool.poetry.dev-dependencies]
pyyaml = "^6.0.1"
httpx = "^0.24.1"
pytest = "^7.4.1"
pytest-azurepipelines = "^1.0.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
